+++ rusqlite/libsqlite3-sys/build.rs
@@ -4,17 +4,21 @@ fn main() {
 
 #[cfg(feature = "bundled")]
 mod build {
-    extern crate gcc;
+    extern crate cc;
     use std::{env, fs};
     use std::path::Path;
 
     pub fn main() {
+        if cfg!(feature = "sqlcipher") {
+            panic!("Builds with bundled SQLCipher are not supported");
+        }
+
         let out_dir = env::var("OUT_DIR").unwrap();
         let out_path = Path::new(&out_dir).join("bindgen.rs");
         fs::copy("sqlite3/bindgen_bundled_version.rs", out_path)
             .expect("Could not copy bindings to output directory");
 
-        gcc::Config::new()
+        cc::Build::new()
             .file("sqlite3/sqlite3.c")
             .flag("-DSQLITE_CORE")
             .flag("-DSQLITE_DEFAULT_FOREIGN_KEYS=1")
@@ -34,6 +38,7 @@ mod build {
             .flag("-DSQLITE_SOUNDEX")
             .flag("-DSQLITE_THREADSAFE=1")
             .flag("-DSQLITE_USE_URI")
+            .flag("-DHAVE_USLEEP=1")
             .compile("libsqlite3.a");
     }
 }
@@ -41,6 +46,9 @@ mod build {
 #[cfg(not(feature = "bundled"))]
 mod build {
     extern crate pkg_config;
+
+    #[cfg(all(feature = "vcpkg", target_env = "msvc"))]
+    extern crate vcpkg;
 
     use std::env;
 
@@ -54,8 +62,9 @@ mod build {
         fn from(header: HeaderLocation) -> String {
             match header {
                 HeaderLocation::FromEnvironment => {
-                    let mut header = env::var("SQLITE3_INCLUDE_DIR")
-                        .expect("SQLITE3_INCLUDE_DIR must be set if SQLITE3_LIB_DIR is set");
+                    let prefix = env_prefix();
+                    let mut header = env::var(format!("{}_INCLUDE_DIR", prefix))
+                        .expect(&format!("{}_INCLUDE_DIR must be set if {}_LIB_DIR is set", prefix, prefix));
                     header.push_str("/sqlite3.h");
                     header
                 }
@@ -72,15 +81,21 @@ mod build {
 
     // Prints the necessary cargo link commands and returns the path to the header.
     fn find_sqlite() -> HeaderLocation {
+        let link_lib = link_lib();
+
         // Allow users to specify where to find SQLite.
-        if let Ok(dir) = env::var("SQLITE3_LIB_DIR") {
-            println!("cargo:rustc-link-lib=sqlite3");
+        if let Ok(dir) = env::var(format!("{}_LIB_DIR", env_prefix())) {
+            println!("cargo:rustc-link-lib={}", link_lib);
             println!("cargo:rustc-link-search={}", dir);
             return HeaderLocation::FromEnvironment;
         }
 
+        if let Some(header) = try_vcpkg() {
+            return header;
+        }
+
         // See if pkg-config can do everything for us.
-        match pkg_config::Config::new().print_system_libs(false).probe("sqlite3") {
+        match pkg_config::Config::new().print_system_libs(false).probe(link_lib) {
             Ok(mut lib) => {
                 if let Some(mut header) = lib.include_paths.pop() {
                     header.push("sqlite3.h");
@@ -94,9 +109,42 @@ mod build {
                 // request and hope that the library exists on the system paths. We used to
                 // output /usr/lib explicitly, but that can introduce other linking problems; see
                 // https://github.com/jgallagher/rusqlite/issues/207.
-                println!("cargo:rustc-link-lib=sqlite3");
+                println!("cargo:rustc-link-lib={}", link_lib);
                 HeaderLocation::Wrapper
             }
+        }
+    }
+
+    #[cfg(all(feature = "vcpkg", target_env = "msvc"))]
+    fn try_vcpkg() -> Option<HeaderLocation> {
+        // See if vcpkg can find it.
+        if let Ok(mut lib) = vcpkg::Config::new().probe(link_lib()) {
+            if let Some(mut header) = lib.include_paths.pop() {
+                header.push("sqlite3.h");
+                return Some(HeaderLocation::FromPath(header.to_string_lossy().into()));
+            }
+        }
+        None
+    }
+
+    #[cfg(not(all(feature = "vcpkg", target_env = "msvc")))]
+    fn try_vcpkg() -> Option<HeaderLocation> {
+        None
+    }
+
+    fn env_prefix() -> &'static str {
+        if cfg!(feature = "sqlcipher") {
+            "SQLCIPHER"
+        } else {
+            "SQLITE3"
+        }
+    }
+
+    fn link_lib() -> &'static str {
+        if cfg!(feature = "sqlcipher") {
+            "sqlcipher"
+        } else {
+            "sqlite3"
         }
     }
 
@@ -139,7 +187,7 @@ mod build {
     mod bindings {
         extern crate bindgen;
 
-        use self::bindgen::chooser::{TypeChooser, IntKind};
+        use self::bindgen::callbacks::{ParseCallbacks, IntKind};
         use super::HeaderLocation;
 
         use std::env;
@@ -150,7 +198,7 @@ mod build {
         #[derive(Debug)]
         struct SqliteTypeChooser;
 
-        impl TypeChooser for SqliteTypeChooser {
+        impl ParseCallbacks for SqliteTypeChooser {
             fn int_macro(&self, _name: &str, value: i64) -> Option<IntKind> {
                 if value >= i32::min_value() as i64 && value <= i32::max_value() as i64 {
                     Some(IntKind::I32)
@@ -166,7 +214,7 @@ mod build {
             let mut output = Vec::new();
             bindgen::builder()
                 .header(header.clone())
-                .type_chooser(Box::new(SqliteTypeChooser))
+                .parse_callbacks(Box::new(SqliteTypeChooser))
                 .generate()
                 .expect(&format!("could not run bindgen on header {}", header))
                 .write(Box::new(&mut output))
@@ -177,7 +225,7 @@ mod build {
             // was added in SQLite 3.8.3, but oring it in in prior versions of SQLite is harmless. We
             // don't want to not build just because this flag is missing (e.g., if we're linking against
             // SQLite 3.7.x), so append the flag manually if it isn't present in bindgen's output.
-            if !output.contains("pub const SQLITE_DETERMINISTIC:") {
+            if !output.contains("pub const SQLITE_DETERMINISTIC") {
                 output.push_str("\npub const SQLITE_DETERMINISTIC: i32 = 2048;\n");
             }