+++ rusqlite/src/load_extension_guard.rs
@@ -25,7 +25,8 @@ impl<'conn> LoadExtensionGuard<'conn> {
     /// Attempt to enable loading extensions. Loading extensions will be disabled when this
     /// guard goes out of scope. Cannot be meaningfully nested.
     pub fn new(conn: &Connection) -> Result<LoadExtensionGuard> {
-        conn.load_extension_enable().map(|_| LoadExtensionGuard { conn: conn })
+        conn.load_extension_enable()
+            .map(|_| LoadExtensionGuard { conn: conn })
     }
 }