+++ rusqlite/src/types/chrono.rs
@@ -3,7 +3,7 @@ extern crate chrono;
 
 use std::borrow::Cow;
 
-use self::chrono::{NaiveDate, NaiveTime, NaiveDateTime, DateTime, TimeZone, UTC, Local};
+use self::chrono::{NaiveDate, NaiveTime, NaiveDateTime, DateTime, TimeZone, Utc, Local};
 
 use Result;
 use types::{FromSql, FromSqlError, FromSqlResult, ToSql, ToSqlOutput, ValueRef};
@@ -19,10 +19,12 @@ impl ToSql for NaiveDate {
 /// "YYYY-MM-DD" => ISO 8601 calendar date without timezone.
 impl FromSql for NaiveDate {
     fn column_result(value: ValueRef) -> FromSqlResult<Self> {
-        value.as_str().and_then(|s| match NaiveDate::parse_from_str(s, "%Y-%m-%d") {
-            Ok(dt) => Ok(dt),
-            Err(err) => Err(FromSqlError::Other(Box::new(err))),
-        })
+        value
+            .as_str()
+            .and_then(|s| match NaiveDate::parse_from_str(s, "%Y-%m-%d") {
+                          Ok(dt) => Ok(dt),
+                          Err(err) => Err(FromSqlError::Other(Box::new(err))),
+                      })
     }
 }
 
@@ -37,17 +39,19 @@ impl ToSql for NaiveTime {
 /// "HH:MM"/"HH:MM:SS"/"HH:MM:SS.SSS" => ISO 8601 time without timezone.
 impl FromSql for NaiveTime {
     fn column_result(value: ValueRef) -> FromSqlResult<Self> {
-        value.as_str().and_then(|s| {
-            let fmt = match s.len() {
-                5 => "%H:%M",
-                8 => "%H:%M:%S",
-                _ => "%H:%M:%S%.f",
-            };
-            match NaiveTime::parse_from_str(s, fmt) {
-                Ok(dt) => Ok(dt),
-                Err(err) => Err(FromSqlError::Other(Box::new(err))),
-            }
-        })
+        value
+            .as_str()
+            .and_then(|s| {
+                let fmt = match s.len() {
+                    5 => "%H:%M",
+                    8 => "%H:%M:%S",
+                    _ => "%H:%M:%S%.f",
+                };
+                match NaiveTime::parse_from_str(s, fmt) {
+                    Ok(dt) => Ok(dt),
+                    Err(err) => Err(FromSqlError::Other(Box::new(err))),
+                }
+            })
     }
 }
 
@@ -63,30 +67,32 @@ impl ToSql for NaiveDateTime {
 /// without timezone. ("YYYY-MM-DDTHH:MM:SS"/"YYYY-MM-DDTHH:MM:SS.SSS" also supported)
 impl FromSql for NaiveDateTime {
     fn column_result(value: ValueRef) -> FromSqlResult<Self> {
-        value.as_str().and_then(|s| {
-            let fmt = if s.len() >= 11 && s.as_bytes()[10] == b'T' {
-                "%Y-%m-%dT%H:%M:%S%.f"
-            } else {
-                "%Y-%m-%d %H:%M:%S%.f"
-            };
+        value
+            .as_str()
+            .and_then(|s| {
+                let fmt = if s.len() >= 11 && s.as_bytes()[10] == b'T' {
+                    "%Y-%m-%dT%H:%M:%S%.f"
+                } else {
+                    "%Y-%m-%d %H:%M:%S%.f"
+                };
 
-            match NaiveDateTime::parse_from_str(s, fmt) {
-                Ok(dt) => Ok(dt),
-                Err(err) => Err(FromSqlError::Other(Box::new(err))),
-            }
-        })
+                match NaiveDateTime::parse_from_str(s, fmt) {
+                    Ok(dt) => Ok(dt),
+                    Err(err) => Err(FromSqlError::Other(Box::new(err))),
+                }
+            })
     }
 }
 
 /// Date and time with time zone => UTC RFC3339 timestamp ("YYYY-MM-DDTHH:MM:SS.SSS+00:00").
 impl<Tz: TimeZone> ToSql for DateTime<Tz> {
     fn to_sql(&self) -> Result<ToSqlOutput> {
-        Ok(ToSqlOutput::from(self.with_timezone(&UTC).to_rfc3339()))
+        Ok(ToSqlOutput::from(self.with_timezone(&Utc).to_rfc3339()))
     }
 }
 
-/// RFC3339 ("YYYY-MM-DDTHH:MM:SS.SSS[+-]HH:MM") into DateTime<UTC>.
-impl FromSql for DateTime<UTC> {
+/// RFC3339 ("YYYY-MM-DDTHH:MM:SS.SSS[+-]HH:MM") into `DateTime<Utc>`.
+impl FromSql for DateTime<Utc> {
     fn column_result(value: ValueRef) -> FromSqlResult<Self> {
         {
             // Try to parse value as rfc3339 first.
@@ -105,19 +111,19 @@ impl FromSql for DateTime<UTC> {
             };
 
             if let Ok(dt) = DateTime::parse_from_rfc3339(&s) {
-                return Ok(dt.with_timezone(&UTC));
+                return Ok(dt.with_timezone(&Utc));
             }
         }
 
         // Couldn't parse as rfc3339 - fall back to NaiveDateTime.
-        NaiveDateTime::column_result(value).map(|dt| UTC.from_utc_datetime(&dt))
+        NaiveDateTime::column_result(value).map(|dt| Utc.from_utc_datetime(&dt))
     }
 }
 
-/// RFC3339 ("YYYY-MM-DDTHH:MM:SS.SSS[+-]HH:MM") into DateTime<Local>.
+/// RFC3339 ("YYYY-MM-DDTHH:MM:SS.SSS[+-]HH:MM") into `DateTime<Local>`.
 impl FromSql for DateTime<Local> {
     fn column_result(value: ValueRef) -> FromSqlResult<Self> {
-        let utc_dt = try!(DateTime::<UTC>::column_result(value));
+        let utc_dt = try!(DateTime::<Utc>::column_result(value));
         Ok(utc_dt.with_timezone(&Local))
     }
 }
@@ -125,12 +131,13 @@ impl FromSql for DateTime<Local> {
 #[cfg(test)]
 mod test {
     use Connection;
-    use super::chrono::{DateTime, Local, NaiveDate, NaiveDateTime, NaiveTime, TimeZone, UTC,
+    use super::chrono::{DateTime, Local, NaiveDate, NaiveDateTime, NaiveTime, TimeZone, Utc,
                         Duration};
 
     fn checked_memory_handle() -> Connection {
         let db = Connection::open_in_memory().unwrap();
-        db.execute_batch("CREATE TABLE foo (t TEXT, i INTEGER, f FLOAT, b BLOB)").unwrap();
+        db.execute_batch("CREATE TABLE foo (t TEXT, i INTEGER, f FLOAT, b BLOB)")
+            .unwrap();
         db
     }
 
@@ -138,11 +145,14 @@ mod test {
     fn test_naive_date() {
         let db = checked_memory_handle();
         let date = NaiveDate::from_ymd(2016, 2, 23);
-        db.execute("INSERT INTO foo (t) VALUES (?)", &[&date]).unwrap();
+        db.execute("INSERT INTO foo (t) VALUES (?)", &[&date])
+            .unwrap();
 
-        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!("2016-02-23", s);
-        let t: NaiveDate = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let t: NaiveDate = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(date, t);
     }
 
@@ -150,11 +160,14 @@ mod test {
     fn test_naive_time() {
         let db = checked_memory_handle();
         let time = NaiveTime::from_hms(23, 56, 4);
-        db.execute("INSERT INTO foo (t) VALUES (?)", &[&time]).unwrap();
+        db.execute("INSERT INTO foo (t) VALUES (?)", &[&time])
+            .unwrap();
 
-        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!("23:56:04", s);
-        let v: NaiveTime = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let v: NaiveTime = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(time, v);
     }
 
@@ -165,15 +178,20 @@ mod test {
         let time = NaiveTime::from_hms(23, 56, 4);
         let dt = NaiveDateTime::new(date, time);
 
-        db.execute("INSERT INTO foo (t) VALUES (?)", &[&dt]).unwrap();
+        db.execute("INSERT INTO foo (t) VALUES (?)", &[&dt])
+            .unwrap();
 
-        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!("2016-02-23T23:56:04", s);
-        let v: NaiveDateTime = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let v: NaiveDateTime = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(dt, v);
 
-        db.execute("UPDATE foo set b = datetime(t)", &[]).unwrap(); // "YYYY-MM-DD HH:MM:SS"
-        let hms: NaiveDateTime = db.query_row("SELECT b FROM foo", &[], |r| r.get(0)).unwrap();
+        db.execute("UPDATE foo set b = datetime(t)", &[])
+            .unwrap(); // "YYYY-MM-DD HH:MM:SS"
+        let hms: NaiveDateTime = db.query_row("SELECT b FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(dt, hms);
     }
 
@@ -183,26 +201,30 @@ mod test {
         let date = NaiveDate::from_ymd(2016, 2, 23);
         let time = NaiveTime::from_hms_milli(23, 56, 4, 789);
         let dt = NaiveDateTime::new(date, time);
-        let utc = UTC.from_utc_datetime(&dt);
+        let utc = Utc.from_utc_datetime(&dt);
 
-        db.execute("INSERT INTO foo (t) VALUES (?)", &[&utc]).unwrap();
+        db.execute("INSERT INTO foo (t) VALUES (?)", &[&utc])
+            .unwrap();
 
-        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!("2016-02-23T23:56:04.789+00:00", s);
 
-        let v1: DateTime<UTC> = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let v1: DateTime<Utc> = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(utc, v1);
 
-        let v2: DateTime<UTC> = db.query_row("SELECT '2016-02-23 23:56:04.789'", &[], |r| r.get(0))
+        let v2: DateTime<Utc> = db.query_row("SELECT '2016-02-23 23:56:04.789'", &[], |r| r.get(0))
             .unwrap();
         assert_eq!(utc, v2);
 
-        let v3: DateTime<UTC> = db.query_row("SELECT '2016-02-23 23:56:04'", &[], |r| r.get(0))
+        let v3: DateTime<Utc> = db.query_row("SELECT '2016-02-23 23:56:04'", &[], |r| r.get(0))
             .unwrap();
         assert_eq!(utc - Duration::milliseconds(789), v3);
 
-        let v4: DateTime<UTC> =
-            db.query_row("SELECT '2016-02-23 23:56:04.789+00:00'", &[], |r| r.get(0)).unwrap();
+        let v4: DateTime<Utc> =
+            db.query_row("SELECT '2016-02-23 23:56:04.789+00:00'", &[], |r| r.get(0))
+                .unwrap();
         assert_eq!(utc, v4);
     }
 
@@ -214,13 +236,16 @@ mod test {
         let dt = NaiveDateTime::new(date, time);
         let local = Local.from_local_datetime(&dt).single().unwrap();
 
-        db.execute("INSERT INTO foo (t) VALUES (?)", &[&local]).unwrap();
+        db.execute("INSERT INTO foo (t) VALUES (?)", &[&local])
+            .unwrap();
 
         // Stored string should be in UTC
-        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let s: String = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert!(s.ends_with("+00:00"));
 
-        let v: DateTime<Local> = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let v: DateTime<Local> = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(local, v);
     }
 }