+++ rusqlite/src/types/serde_json.rs
@@ -32,7 +32,8 @@ mod test {
 
     fn checked_memory_handle() -> Connection {
         let db = Connection::open_in_memory().unwrap();
-        db.execute_batch("CREATE TABLE foo (t TEXT, b BLOB)").unwrap();
+        db.execute_batch("CREATE TABLE foo (t TEXT, b BLOB)")
+            .unwrap();
         db
     }
 
@@ -46,9 +47,11 @@ mod test {
                      &[&data, &json.as_bytes()])
             .unwrap();
 
-        let t: serde_json::Value = db.query_row("SELECT t FROM foo", &[], |r| r.get(0)).unwrap();
+        let t: serde_json::Value = db.query_row("SELECT t FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(data, t);
-        let b: serde_json::Value = db.query_row("SELECT b FROM foo", &[], |r| r.get(0)).unwrap();
+        let b: serde_json::Value = db.query_row("SELECT b FROM foo", &[], |r| r.get(0))
+            .unwrap();
         assert_eq!(data, b);
     }
 }